Now can ingest GP priors manually to different decorrelation parameters
