+++ .gitignore
@@ -137,3 +137,4 @@ dmypy.json
 /juls/
 /gpcheops/test.py
 /gpcheops/test1.py
+gpcheops/test2.py
