+++ gpcheops/__init__.py
@@ -1,1 +1,2 @@
 from gpcheops import *
+from utils import *
